Script that bootstraps the machine it's run on.
This is called by bin/bootstrap-server.sh after having been copied to
the machine. Its purpose is to do just enough so that sbtf-bootstrap.sh
can be run - which means very basic environment configuration (locales,
timezone), installation of git to clone the repository, and the set up
of the sbtf user.