Add script to bootstrap an SBTF server environment.
This script runs assuming the environment is ready to be bootstrapped -
a job which either people do themselves locally (e.g. build a
container), or that production-bootstrap.sh does.

With these things done, this script ensures puppet is installed, writes
out configuration to tell puppet what it should be doing, and runs it
for the first time.