Replace application/config/config.php with a slightly briefer one.
The main aim here is to have it in puppet so I can make wholesale
settings changes, to the logging in particular, but also to other
settings.

Have hardcoded the site_domain for now. I'll have to think about how to
include this information in the setup process so there's no ambiguity
(and so we can set the nginx server_name for example).